rename inspector to inspecteur and shedule to schedule
