+++ client/src/views/admin/components/GenerateInspecteurBordereaux.vue
