+++ server/src/routes/business/mail/body-schedule-inspecteur-template.js
@@ -1,4 +1,4 @@
-export const getSheduleInspecteurTemplate = (
+export const getScheduleInspecteurTemplate = (
   inspecteurName,
   inspecteurMatricule,
   dateExam,
