+++ server/src/routes/business/mail/failed-mail-schelude-inspecteurs-template.js
@@ -1,4 +1,4 @@
-export const getFailedSheduleInspecteurTemplate = (date, inspecteurs) => {
+export const getFailedScheduleInspecteurTemplate = (date, inspecteurs) => {
   const list = inspecteurs.reduce((acc, insp) => acc + '<li>' + insp + '</li>')
   return `
     <p> Le mail pour le planning des inspecteurs du ${date} n'a pu être envoyé pour les inspecteurs ci-dessous: </p>
